--- app/static/json/testePagina.json
@@ -1,35 +1,0 @@
-[{
-        "id": 1,
-        "pergunta": "Nome completo do usuário",
-        "opcoes": null,
-        "obrigatoria": true,
-        "tipo": "text",
-        "aba": "Informações basicas",
-        "proxima": 2,
-        "extra": "preencher antes de iniciar o atendimento",
-        "condicao": [3]
-    },
-    {
-        "id": 2,
-        "pergunta": "CPF do usuário",
-        "opcoes": null,
-        "obrigatoria": true,
-        "tipo": "texto",
-        "aba": "Informações basicas",
-        "proxima": 3,
-        "extra": "preencher antes de iniciar o atendimento",
-        "condicao": null
-    },
-    {
-        "id": 3,
-        "pergunta": "Conseguiu realizar o atendimento?",
-        "opcoes": null,
-        "obrigatoria": true,
-        "tipo": "radio",
-        "aba": "Informações basicas",
-        "proxima": 4,
-        "extra": "preencher antes de iniciar o atendimento",
-        "condicao": null
-    }
-
-]
