+++ app/templates/form/form.html
@@ -1,74 +1,74 @@
-{% from 'form/text.html' import text %}
-{% from 'form/radio.html' import radio %}
+{% from 'form/text.html' import text with context %}
+{% from 'form/radio.html' import radio with context %}
 
 
 {% extends "template.html" %}
 {% block conteudo %}
 
-    <div class="row">
-        <div class="col-lg-12">
-            <div class="ibox">
-                <div class="ibox-title">
-                    <h3 style="font-weight: 400;">Formulário Primeiro Atendimento</h3>
-                    <p>Este é o formulário a ser preenchido no primeiro contato telefônico com o usuário</p>
-                </div>
-                <div class="ibox-content p-0 p-md-4">
-                    <form id="form" action="#">
-                        {% for fieldset in fieldsets %}
-                            <h1>{{ fieldset.name }}</h1>
-                            <fieldset class="w-100 p-0 p-md-4">
+<div class="row">
+    <div class="col-lg-12">
+        <div class="ibox">
+            <div class="ibox-title">
+                <h3 style="font-weight: 400;">Formulário Primeiro Atendimento</h3>
+                <p>Este é o formulário a ser preenchido no primeiro contato telefônico com o usuário</p>
+            </div>
+            <div class="ibox-content p-0 p-md-4">
+                <form id="form" action="#">
+                    {% for fieldset in fieldsets %}
+                    <h1>{{ fieldset.name }}</h1>
+                    <fieldset class="w-100 p-0 p-md-4">
+                        <div class="row">
+                            {% for section in fieldset.sections %}
+                            <div class="col-12 col-md-6">
+                                <h3 style="font-weight: 400;">{{ section.name }}</h3>
+                                <hr>
                                 <div class="row">
-                                    {% for section in fieldset.sections %}
-                                        <div class="col-12 col-md-6">
-                                            <h3 style="font-weight: 400;">{{ section.name }}</h3>
-                                            <hr>
-                                            <div class="row">
-                                                {% for input in section.inputs %}
-                                                    <div class="col-12 col-md-6">
-                                                        <div class="form-group">
-                                                            <label>{{ input.label }} {{ '*' if input.required }}</label>
+                                    {% for input in section.inputs %}
+                                    <div class="col-12 col-md-6">
+                                        <div class="form-group">
+                                            <label>{{ input.label }} {{ '*' if input.required }}</label>
 
-                                                            {% if input.type == "radio" %}
-                                                                {{ radio(input) }}
-                                                            {% else %}
-                                                                {{ text(input) }}
-                                                            {% endif %}
+                                            {% if input.type == "radio" %}
+                                            {{ radio(input) }}
+                                            {% else %}
+                                            {{ text(input) }}
+                                            {% endif %}
 
-                                                        </div>
-                                                    </div>
-                                                {% endfor %}
-                                            </div>
                                         </div>
+                                    </div>
                                     {% endfor %}
                                 </div>
-                            </fieldset>
-                        {% endfor %}
-                    </form>
-                </div>
+                            </div>
+                            {% endfor %}
+                        </div>
+                    </fieldset>
+                    {% endfor %}
+                </form>
             </div>
         </div>
     </div>
+</div>
 
 {% endblock %}
 
 {% block script %}
-    <script>
-        function radio(input) {
-            const contentId = 'id-' + input.getAttribute("name")
+<script>
+    function radio(input) {
+        const contentId = 'id-' + input.getAttribute("name")
 
-            if (input.value == 0) {
-                document.getElementById(contentId).style.visibility = "visible"
-            } else {
-                document.getElementById(contentId).style.visibility = "hidden"
-            }
+        if (input.value == 0) {
+            document.getElementById(contentId).style.visibility = "visible"
+        } else {
+            document.getElementById(contentId).style.visibility = "hidden"
         }
+    }
 
-        window.addEventListener('load', function () {
-            document.querySelectorAll(".iCheck-helper").forEach(function (input) {
-                input.addEventListener("click", function (event) {
-                    radio(event.target.parentElement.childNodes[0])
-                })
+    window.addEventListener('load', function () {
+        document.querySelectorAll(".iCheck-helper").forEach(function (input) {
+            input.addEventListener("click", function (event) {
+                radio(event.target.parentElement.childNodes[0])
             })
         })
-    </script>
+    })
+</script>
 {% endblock %}
