+++ app/templates/form/radio.html
@@ -1,26 +1,26 @@
 {% from 'form/text.html' import text %}
 
 {% macro radio(input) -%}
-    {% for option in input.options %}
-        <div class="i-checks">
-            <label onclick="radio(this.childNodes[1].childNodes[0])">
-                <input type="radio" checked="" value="{{ loop.index0 }}" name="{{ input.name }}">
-                <i></i>
-                {{ option }}
-            </label>
-        </div>
+{% for option in input.options %}
+<div class="i-checks">
+    <label onclick="radio(this.childNodes[1].childNodes[0])">
+        <input type="radio" checked="" value="{{ loop.index0 }}" name="{{ input.name }}">
+        <i></i>
+        {{ option }}
+    </label>
+</div>
+{% endfor %}
+
+<div id="id-{{ input.name }}" style="visibility: hidden;">
+    {% for field in input.fields %}
+    <label for="{{ field.name }}">{{ field.label }}</label>
+
+    {% if field.type == "radio" %}
+    {{ radio(field) }}
+    {% else %}
+    {{ text(field) }}
+    {% endif %}
     {% endfor %}
-
-    <div id="id-{{ input.name }}" style="visibility: hidden;">
-        {% for field in input.fields %}
-            <label for="{{ field.name }}">{{ field.label }}</label>
-
-            {% if field.type == "radio" %}
-                {{ radio(field) }}
-            {% else %}
-                {{ text(field) }}
-            {% endif %}
-        {% endfor %}
-    </div>
+</div>
 
 {%- endmacro %}
