+++ app/templates/form/text.html
@@ -1,7 +1,6 @@
 {% macro text(input) -%}
 
-    <input id="{{ input.name }}" name="{{ input.name }}" type="{{ input.type if input.type else 'text' }}"
-           class="form-control {{ 'required' if input.required }} {{ input.class }}"
-           placeholder="{{ input.placeholder }}">
+<input id="{{ input.name }}" name="{{ input.name }}" type="{{ input.type if input.type else 'text' }}"
+    class="form-control {{ 'required' if input.required }} {{ input.class }}" placeholder="{{ input.placeholder }}">
 
 {%- endmacro %}
