--- app/templates/tesjason.html
@@ -1,10 +1,0 @@
-{% extends "template.html" %}
-{% block conteudo %}
-
-{% for pergunta in perguntas %}
-
-{{pergunta.pergunta}}
-
-{% endfor %}
-
-{% endblock %}
